module: split getFormat into its own file
