+++ lib/internal/main/check_syntax.js
@@ -51,13 +51,10 @@ function checkSyntax(source, filename) {
   if (filename === '[stdin]' || filename === '[eval]') {
     isModule = getOptionValue('--input-type') === 'module';
   } else {
-    const {
-      defaultResolve: resolve,
-      defaultGetFormat: getFormat
-    } = require('internal/modules/esm/default_resolve');
-    const { url } = resolve(
-      pathToFileURL(filename).toString());
-    const { format } = getFormat(url);
+    const { defaultResolve } = require('internal/modules/esm/resolve');
+    const { defaultGetFormat } = require('internal/modules/esm/get_format');
+    const { url } = defaultResolve(pathToFileURL(filename).toString());
+    const { format } = defaultGetFormat(url);
     isModule = format === 'module';
   }
   if (isModule) {
