+++ lib/internal/modules/esm/loader.js
@@ -18,10 +18,8 @@ const { validateString } = require('internal/validators');
 const ModuleMap = require('internal/modules/esm/module_map');
 const ModuleJob = require('internal/modules/esm/module_job');
 
-const {
-  defaultResolve,
-  defaultGetFormat
-} = require('internal/modules/esm/default_resolve');
+const { defaultResolve } = require('internal/modules/esm/resolve');
+const { defaultGetFormat } = require('internal/modules/esm/get_format');
 const createDynamicModule = require(
   'internal/modules/esm/create_dynamic_module');
 const { defaultGetSource, defaultTransformSource, translators } = require(
